refactor: adapterのディレクトリを移動
